refactor: Integrate expo lib with universal interface
Only outstanding caveat is that grabbing the right fingerprint in cuid 
does not work when bundling this way, so it attempts to grab userAgent 
when generating a client ID, which does not exist. Rectifying that one 
line manually gives you a working Expo client!

Resolving this requires somehow grabbing the react-native variant of
cuid/fingerprint?